+++ index.js
@@ -1,9 +1,9 @@
+const knex = require('knex');
 const postgres = require('./src/postgres');
 const mysql = require('./src/mysql');
-const knex = require('knex');
 
 module.exports = {
   postgres,
   mysql,
-  knex,
+  knex
 };
