+++ src/helper/sqlstring.js
