+++ src/mysql.js
@@ -1,5 +1,5 @@
 const MySqlClient = require('knex/lib/dialects/mysql');
-const dataApi = require('./data-api');
+const knexAdapter = require('./knex-adapter');
 const constants = require('./constants');
 
 // Call mysql client to setup knex, this set as this function
@@ -9,6 +9,6 @@ const client = MySqlClient.constructor
       MySqlClient.call(this, config);
     };
 
-dataApi(client, MySqlClient, constants.dialects.mysql);
+knexAdapter(client, MySqlClient, constants.dialects.mysql);
 
 module.exports = client;
